+++ krb5/src/windows/lib/registry.c
@@ -4,6 +4,7 @@
  * Author:  Michael Graff
  */
 
+#include <stdlib.h>
 #include <windows.h>
 #include <windowsx.h>
 
@@ -115,7 +116,7 @@ registry_string_set(HKEY hkey, char *sub, char *x)
 {
 	DWORD   err;
 
-	err = RegSetValueEx(hkey, sub, 0, REG_SZ, (BYTE *)x, strlen(x) + 1);
+	err = RegSetValueEx(hkey, sub, 0, REG_SZ, (BYTE *)x, (DWORD)strlen(x) + 1);
 	if (err != ERROR_SUCCESS)
 		return -1;
